# Conflicts:
#	do-not-forget-backend/src/main/java/pl/pw/as/rest/controller/TaskController.java
#	do-not-forget-backend/src/main/java/pl/pw/as/services/TaskService.java
#	do-not-forget-frontend/src/main/resources/webapp/src/services/TaskService.ts
#	do-not-forget-frontend/src/main/resources/webapp/src/services/config.ts

+++ do-not-forget-frontend/src/main/resources/webapp/src/services/ErrorService.ts
@@ -7,6 +7,12 @@ import {LOGIN} from "../navigation.constants";
 export class ErrorService {
     private static RUNTIME_EXCEPTION_ID: string = 'java.lang.RuntimeException';
     private static JWT_EXPIRED_EXCEPTION_ID: string = 'io.jsonwebtoken.ExpiredJwtException';
+    private static WRONG_CREDENTIALS_ERROR_DATA = {
+        error: "Unauthorized",
+        message: "Authentication Failed: Bad credentials",
+        path: "/login",
+        status: 401
+    };
 
     constructor(private _router: Router) {
     }
@@ -15,16 +21,22 @@ export class ErrorService {
     handleExceptionAndReturnMessage(error): string {
         let exception = JSON.parse(error._body);
 
+        if (this._isBadCredentials(exception)) {
+            return 'Wrong login or password, try again';
+        }
+
         if (this._isRuntimeException(exception)) {
             return exception.message;
 
-        } else if (this._isTokenExpired(exception)) {
+        }
+
+        if (this._isTokenExpired(exception)) {
             sessionStorage.clear();
             this._router.navigate([LOGIN, LoginStateMapService.TOKEN_EXPIRED_STATE]);
             return '';
-
         }
-        return 'Internal server error, please contact support.';
+
+        return 'Something went wrong, please try again later.';
     }
 
     private _isRuntimeException(exception) {
@@ -35,5 +47,12 @@ export class ErrorService {
     private _isTokenExpired(exception: any) {
         return exception.status == 500
             && exception.exception == ErrorService.JWT_EXPIRED_EXCEPTION_ID;
+    }
+
+    private _isBadCredentials(exception: any) {
+        return exception.status == ErrorService.WRONG_CREDENTIALS_ERROR_DATA.status
+            && exception.error == ErrorService.WRONG_CREDENTIALS_ERROR_DATA.error
+            && exception.path == ErrorService.WRONG_CREDENTIALS_ERROR_DATA.path
+            && exception.message == ErrorService.WRONG_CREDENTIALS_ERROR_DATA.message;
     }
 }
